rename functions to utilities
