+++ packages/runed/src/lib/index.ts
@@ -1,1 +1,1 @@
-export * from "./functions/index.js";
+export * from "./utilities/index.js";
