+++ packages/runed/src/lib/utilities/IsSupported/IsSupported.ts
@@ -3,7 +3,7 @@
  *
  * Useful for checking if a browser API is supported before attempting to use it.
  *
- * @see {@link https://runed.dev/docs/functions/use-supported}
+ * @see {@link https://runed.dev/docs/utilities/use-supported}
  */
 export class IsSupported {
 	#current = $state(false);
