+++ packages/runed/src/lib/utilities/Previous/Previous.svelte.ts
@@ -4,7 +4,7 @@ import type { Getter } from "$lib/internal/types.js";
 /**
  * Holds the previous value of a getter.
  *
- * @see {@link https://runed.dev/docs/functions/use-previous}
+ * @see {@link https://runed.dev/docs/utilities/use-previous}
  */
 export class Previous<T> {
 	#previous = $state<T | undefined>(undefined)
