+++ packages/runed/src/lib/utilities/activeElement/activeElement.svelte.ts
@@ -4,7 +4,7 @@ import { readable } from "../readable/readable.svelte.js";
  * An object holding a reactive value that is equal to `document.activeElement`.
  * It automatically listens for changes, keeping the reference up to date.
  *
- * @see {@link https://runed.dev/docs/functions/use-active-element}
+ * @see {@link https://runed.dev/docs/utilities/use-active-element}
  */
 export const activeElement = readable(
 	() => document.activeElement,
